Use requireFile builder for pragmata-pro
This commit updates the pragmata-pro package to use the
requireFile builder from nixpkgs. This provides a clearer
error when the required files are not present.

Also, I got sick of doing `git add -Nf` so that it would be added
to this flake. The explanations I found on the web often failed to
discuss any of these additional complications.